# docs/api/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "카페 프로젝트 명세서",
    "version": "0.0.1",
    "title": "Cafe Project API",
    "contact": {
      "name": "김용환",
      "url": "https://yonghwankim-dev.tistory.com/",
      "email": "yh.kim951107@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 리스트 조회",
        "operationId": "listQuestionUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호",
            "required": false,
            "type": "string",
            "default": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/qna": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 리스트 조회",
        "operationId": "listQuestionUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호",
            "required": false,
            "type": "string",
            "default": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 추가",
        "operationId": "addQuestionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "questionRequest",
            "description": "질문 추가 저장 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionSavedRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/qna/new": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 추가 페이지",
        "operationId": "addQuestionFormUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/qna/{id}": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 조회",
        "operationId": "detailQuestionUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "댓글 커서",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "게시글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 수정",
        "operationId": "modifyQuestionUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "수정 게시글 등록번호",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionRequest",
            "description": "수정 게시글 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionSavedRequest"
            }
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보가 담긴 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 삭제",
        "operationId": "deleteQuestionUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "삭제 게시글 등록번호",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보가 담긴 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/qna/{id}/edit": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 수정 페이지",
        "operationId": "editQuestionFormUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "수정 게시글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "전체 회원을 조회하는 메소드",
        "operationId": "listUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "회원을 생성하는 메소드",
        "operationId": "createUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userRequest",
            "description": "생성할 회원의 정보를 담은 객체",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSavedRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/new": {
      "get": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "회원 가입 페이지",
        "operationId": "addUserFormUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "특정한 회원을 조회하는 메소드",
        "operationId": "detailUserUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "조회하고자 하는 회원 아이디",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "회원 정보를 수정하는 메소드",
        "operationId": "modifyUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "수정하고자 하는 회원 등록번호",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보가 담긴 세션",
            "required": false
          },
          {
            "in": "query",
            "name": "userRequest",
            "description": "수정된 회원 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserSavedRequest"
            }
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/{id}/edit": {
      "get": {
        "tags": [
          "사용자 API 정보를 제공하는 Controller"
        ],
        "summary": "회원 수정 페이지",
        "operationId": "modifyUserFormUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "수정하고자 하는 회원 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ModelAndView": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "boolean"
        },
        "model": {
          "type": "object"
        },
        "modelMap": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "reference": {
          "type": "boolean"
        },
        "status": {
          "type": "string",
          "enum": [
            "ACCEPTED",
            "ALREADY_REPORTED",
            "BAD_GATEWAY",
            "BAD_REQUEST",
            "BANDWIDTH_LIMIT_EXCEEDED",
            "CHECKPOINT",
            "CONFLICT",
            "CONTINUE",
            "CREATED",
            "DESTINATION_LOCKED",
            "EXPECTATION_FAILED",
            "FAILED_DEPENDENCY",
            "FORBIDDEN",
            "FOUND",
            "GATEWAY_TIMEOUT",
            "GONE",
            "HTTP_VERSION_NOT_SUPPORTED",
            "IM_USED",
            "INSUFFICIENT_SPACE_ON_RESOURCE",
            "INSUFFICIENT_STORAGE",
            "INTERNAL_SERVER_ERROR",
            "I_AM_A_TEAPOT",
            "LENGTH_REQUIRED",
            "LOCKED",
            "LOOP_DETECTED",
            "METHOD_FAILURE",
            "METHOD_NOT_ALLOWED",
            "MOVED_PERMANENTLY",
            "MOVED_TEMPORARILY",
            "MULTIPLE_CHOICES",
            "MULTI_STATUS",
            "NETWORK_AUTHENTICATION_REQUIRED",
            "NON_AUTHORITATIVE_INFORMATION",
            "NOT_ACCEPTABLE",
            "NOT_EXTENDED",
            "NOT_FOUND",
            "NOT_IMPLEMENTED",
            "NOT_MODIFIED",
            "NO_CONTENT",
            "OK",
            "PARTIAL_CONTENT",
            "PAYLOAD_TOO_LARGE",
            "PAYMENT_REQUIRED",
            "PERMANENT_REDIRECT",
            "PRECONDITION_FAILED",
            "PRECONDITION_REQUIRED",
            "PROCESSING",
            "PROXY_AUTHENTICATION_REQUIRED",
            "REQUESTED_RANGE_NOT_SATISFIABLE",
            "REQUEST_ENTITY_TOO_LARGE",
            "REQUEST_HEADER_FIELDS_TOO_LARGE",
            "REQUEST_TIMEOUT",
            "REQUEST_URI_TOO_LONG",
            "RESET_CONTENT",
            "SEE_OTHER",
            "SERVICE_UNAVAILABLE",
            "SWITCHING_PROTOCOLS",
            "TEMPORARY_REDIRECT",
            "TOO_EARLY",
            "TOO_MANY_REQUESTS",
            "UNAUTHORIZED",
            "UNAVAILABLE_FOR_LEGAL_REASONS",
            "UNPROCESSABLE_ENTITY",
            "UNSUPPORTED_MEDIA_TYPE",
            "UPGRADE_REQUIRED",
            "URI_TOO_LONG",
            "USE_PROXY",
            "VARIANT_ALSO_NEGOTIATES"
          ]
        },
        "view": {
          "$ref": "#/definitions/View"
        },
        "viewName": {
          "type": "string"
        }
      },
      "title": "ModelAndView"
    },
    "QuestionResponse": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "writer": {
          "type": "string"
        }
      },
      "title": "QuestionResponse"
    },
    "QuestionSavedRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "title": {
          "type": "string",
          "pattern": "^.{1,100}$"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "QuestionSavedRequest"
    },
...
    "UserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "UserResponse"
    },
    "UserSavedRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "pattern": "[a-z\\d]+@[a-z]+\\.[a-z]{2,3}"
        },
        "name": {
          "type": "string",
          "pattern": "^[a-z가-힣]{1,20}$"
        },
        "password": {
          "type": "string",
          "pattern": "[A-Za-z\\d!@#$%^&*()_+]{8,16}$"
        },
        "userId": {
          "type": "string",
          "pattern": "^[a-z\\d_-]{5,20}$"
        }
      },
      "title": "UserSavedRequest"
    },
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}